backend/requirements.txt
numpy>=1.26
sounddevice>=0.4.6
soundfile>=0.12.1
pyyaml>=6.0.1
websockets>=12.0
openwakeword>=0.6.0
silero-vad>=5.1
faster-whisper>=1.0.3
resemblyzer>=0.1.4
piper-tts>=1.2.0
ollama>=0.3.3
psutil>=5.9.8
requests>=2.32
keyboard>=0.13.5
pycaw>=20240210; sys_platform == "win32"
pywin32>=306; sys_platform == "win32"
comtypes>=1.4.1; sys_platform == "win32"
mss>=9.0.0
Pillow>=10.0.0
pyautogui>=0.9.54; sys_platform == "win32"
